requirements.txt
beautifulsoup4
colorama==0.4.6
openai==0.27.2
anthropic
google-generativeai
playsound==1.2.2
python-dotenv==1.0.0
pyyaml==6.0
readability-lxml==0.8.1
requests
tiktoken==0.3.3
gTTS==2.3.1
docker
duckduckgo-search==3.5.0
google-api-python-client #(https://developers.google.com/custom-search/v1/overview)
pinecone-client==2.2.1
redis
orjson
Pillow
selenium
webdriver-manager
jsonschema
tweepy
click

##Dev
coverage
flake8
numpy
pre-commit
black
sourcery
isort
gitpython==3.1.31

# Items below this point will not be included in the Docker Image

# Testing dependencies
pytest
asynctest
pytest-asyncio
pytest-benchmark
pytest-cov
pytest-integration
pytest-mock
